memory: changed all code to use new memory architecture

+++ src/cpu/emu_instructions.rs
@@ -231,7 +231,7 @@ impl CPU for EmulationCPU {
         println!("{:<6} {}", "jmp", arg);
         arg.assert_one_argument();
         match arg.first_argument {
-            InstructionArgument::Immediate { immediate } => machine_state.rip += immediate as usize,
+            InstructionArgument::Immediate { immediate } => machine_state.rip += immediate,
             _ => panic!("JMP: Unsupported argument."),
         }
     }
